requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
bcrypt
python-multipart
